backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.1
passlib==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
httpx==0.27.2